app/requirements.txt
fastapi==0.111.0

uvicorn[standard]==0.30.1

pydantic==2.7.4

pydantic-settings==2.3.4

sqlmodel==0.0.21

psycopg[binary]==3.2.1

python-multipart==0.0.9

email-validator==2.2.0

google-api-python-client==2.136.0

google-auth==2.31.0

google-auth-oauthlib==1.2.1

msal==1.31.0

requests==2.32.3

python-dateutil==2.9.0.post0

pytz==2024.1

tzdata==2024.1

passlib[bcrypt]==1.7.4

pyjwt==2.8.0

httpx==0.27.0

tenacity==8.5.0

orjson==3.10.7

